+++ .gitignore
@@ -7,7 +7,7 @@ platforms/
 # NativeScript Template
 *.js.map
 *.js
-!webpack.config.js
+!react-nativescript.webpack.config.js
 !hooks/after-createProject/after-createProject.js
 
 # Logs
