+++ .npmignore
@@ -3,7 +3,7 @@ npm-debug.log
 
 *.js.map
 *.js
-!webpack.config.js
+!react-nativescript.webpack.config.js
 hooks/*
 !hooks/after-createProject/after-createProject.js
 lib/
