+++ nsconfig.json
@@ -1,3 +1,4 @@
 {
-  "useLegacyWorkflow": false
+  "useLegacyWorkflow": false,
+  "webpackConfigPath": "./react-nativescript.webpack.config.js"
 }
